test data deleting dragged to @After method

+++ src/test/java/site/stellarburgers/order/GetOrdersTest.java
@@ -1,19 +1,28 @@
 package site.stellarburgers.order;
 
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 
 @DisplayName("Получение заказов")
 public class GetOrdersTest extends OrderBaseTest {
 
+    String accessToken;
+
+    @After
+    public void deleteUser(){
+        if (accessToken != null) {
+            userClient.delete(accessToken);
+        }
+    }
+
     @Test
     @DisplayName("Получение заказов как авторизованный пользователь")
     public void getOrdersAuthorizedReturnSuccess() {
-        String accessToken = OrderBaseTest.createUserAndOrder(order, orderClient, user, userClient);
+        accessToken = OrderBaseTest.createUserAndOrder(order, orderClient, user, userClient);
         boolean isSuccess = orderClient.getOrdersWithAuth(accessToken);
         Assert.assertTrue(isSuccess);
-        userClient.delete(accessToken);
     }
 
     @Test
